+++ assets/app/app/citizen-scientist/expert/index.js
@@ -55,9 +55,8 @@ angular.module('a2.citizen-scientist.expert', [
     };
 
     $scope.getTemplateVisualizerUrl = function(template){
-        var projecturl = Project.getUrl();
         var box = ['box', template.x1, template.y1, template.x2, template.y2].join(',')
-        return template ? "/visualizer/"+projecturl+"/visualizer/rec/"+template.recording+"?a="+box : '';
+        return template ? "/project/"+template.source_project_uri+"/visualizer/rec/"+template.recording+"?a="+box : '';
     },
 
     $scope.selectItem = function(patternmatchingId){
@@ -270,7 +269,7 @@ angular.module('a2.citizen-scientist.expert', [
 
     getTemplateVisualizerUrl: function(template){
         var box = ['box', template.x1, template.y1, template.x2, template.y2].join(',')
-        return template ? "/visualizer/"+this.projecturl+"/visualizer/rec/"+template.recording+"?a="+box : '';
+        return template ? "/visualizer/"+template.source_project_uri+"/visualizer/rec/"+template.recording+"?a="+box : '';
     },
 
     setRoi: function(roi_index){